Fixed multi user message issues with merkle proofs

+++ src/api/webrtc/disconnectFromChannelLabelQuery.ts
@@ -21,7 +21,10 @@ const webrtcDisconnectFromChannelLabelQuery: BaseQueryFn<
   RTCDisconnectFromChannelLabelParamsExtension,
   void,
   unknown
-> = async ({ label, alsoDeleteData, dataChannels }, api) => {
+> = async (
+  { label, alsoDeleteData, alsoSendFinishedMessage, dataChannels },
+  api,
+) => {
   try {
     const CHANNELS_LEN = dataChannels.length;
     if (CHANNELS_LEN === 0) return { data: undefined };
@@ -34,13 +37,20 @@ const webrtcDisconnectFromChannelLabelQuery: BaseQueryFn<
       )
         continue;
 
-      dataChannels[i].close();
+      if (alsoDeleteData || alsoSendFinishedMessage) {
+        const { hash, merkleRootHex } =
+          await decompileChannelMessageLabel(label);
 
-      if (alsoDeleteData) {
-        const { merkleRootHex } = await decompileChannelMessageLabel(label);
+        if (alsoSendFinishedMessage && dataChannels[i].readyState === "open") {
+          dataChannels[i].send(hash.buffer as ArrayBuffer);
+        }
 
-        api.dispatch(deleteMessage({ merkleRootHex }));
+        if (alsoDeleteData) {
+          api.dispatch(deleteMessage({ merkleRootHex }));
+        }
       }
+
+      dataChannels[i].close();
     }
 
     return { data: undefined };
